sure

+++ app/src/main/java/com/example/telematics_engine/Accelerometers.kt
@@ -1,7 +1,7 @@
 package com.example.telematics_engine
 
 data class Accelerometers(
-    val x:Int=0,
-    val y:Int=0,
-    val z:Int=0,
+    val x:Float=0f,
+    val y:Float=0f,
+    val z:Float=0f,
 )
